pyproject.toml
[project]
name = "andersenlab-todo-list-helper"
version = "0.1.0"
description = "Django REST API for managing personal tasks"
authors = [
    {name = "Timofey Prokofyev",email = "tryatim8@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "django (>=5.2.3,<6.0.0)",
    "djangorestframework (>=3.16.0,<4.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "djangorestframework-simplejwt (>=5.5.0,<6.0.0)",
    "drf-spectacular (>=0.28.0,<0.29.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "django-filter (>=25.1,<26.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pytest = "^8.4.1"
pytest-django = "^4.11.1"
mypy = "^1.16.1"
flake8 = "^7.3.0"
isort = "^6.0.1"
coverage = "^7.9.1"
faker = "^37.4.0"
django-stubs = "^5.2.1"
djangorestframework-stubs = "^3.16.0"

